pyproject.toml
[project]
name = "smartestfactory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.72.0",
    "argparse>=1.4.0",
    "beautifulsoup4>=4.14.2",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.120.4",
    "html2text>=2025.4.15",
    "ipython>=9.6.0",
    "jupyter>=1.1.1",
    "langchain>=1.0.3",
    "numpy>=2.3.4",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "pathlib>=1.0.1",
    "pypdf2>=3.0.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "typing>=3.10.0.0",
    "uvicorn>=0.38.0",
]
